Move SingularModule to swerve folder
